requirements.txt
# Core RAG dependencies
openai==1.35.3
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Discord bot dependencies
discord.py==2.3.2

# Redis dependencies
redis==4.6.0